long double

+++ srcs/events.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   events.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
+/*   By: cmartin- <cmartin-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/26 22:14:59 by cmartin           #+#    #+#             */
-/*   Updated: 2023/11/24 14:40:24 by marvin           ###   ########.fr       */
+/*   Updated: 2023/11/25 14:41:12 by cmartin-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,7 +35,7 @@ void	ft_putdist_wall(t_wallhit distance, t_player *player)
 		player->posy = distance.y;
 }
 
-void	ft_putmov(t_map map, t_player *player, double angle)
+void	ft_putmov(t_map map, t_player *player, long double angle)
 {
 	t_wallhit	distance;
 
@@ -52,7 +52,7 @@ void	ft_putmov(t_map map, t_player *player, double angle)
 
 void	ft_rotation(t_player *player, char c)
 {
-	double	rotation;
+	long double	rotation;
 
 	rotation = M_PI / 30;
 	if (player->angle >= 2 * M_PI)
@@ -70,7 +70,7 @@ void	ft_rotation(t_player *player, char c)
 
 int	ft_mov(t_map map, t_player *player, int key)
 {
-	double	prevangle;
+	long double	prevangle;
 
 	prevangle = player->angle;
 	if (key == 'w')
@@ -104,8 +104,8 @@ int	ft_key(int key, t_data *data)
 		data -> player.angle = 0;
 	if (ft_mov(data->map, &(data->player), key))
 		change_image(data);
-	printf("angle = %f\n", data->player.angle);
-	printf("posx = %f\n", data->player.posx);
-	printf("posy = %f\n", data->player.posy);
+	printf("angle = %Lf\n", data->player.angle);
+	printf("posx = %Lf\n", data->player.posx);
+	printf("posy = %Lf\n", data->player.posy);
 	return (0);
 }
